modify: move single single server codes into servers crate
